Basic listener scheme creating new games
Creates the listener for for the new game buttons 2-5, also dialog box to prompt player for their name at the beginning of the game. Currently storing the game instance and player as currentGame and currentPlayer in the GameGUI class, probably should make a main game class to run everything through. Added getters and setters to the player class.