Looking almost done
Fixed issues with the abstract class and it not ending the 2nd game of the same level. everything seems functional and now on startup you can create instances of game 1.

There was an issue with dark mode when trying to load onto my machine it was pulling a speicific path on aidan's computer.  Try add it to the repo if we are moving with this.

Great work guys coming down the home stretch